requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
httpx==0.27.2
python-dotenv==1.0.1
aiosqlite==0.20.0
pydantic==2.9.2
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.9
fpdf2==2.8.1
markdown==3.7
PyJWT==2.9.0
bleach==6.1.0
prometheus-client==0.21.0
jsonschema==4.23.0
PyYAML==6.0.3
pytest==9.0.2
